+++ os/test/PolicyPrio.cpp
@@ -25,11 +25,11 @@ class Test {
 
 public:
 	static void work() {
-		t1.start(0);
-		t2.start(0);
-		t3.start(1);
-		t4.start(2);
-		t5.start(2);
+		t1.start((uint8_t)0);
+		t2.start((uint8_t)0);
+		t3.start((uint8_t)1);
+		t4.start((uint8_t)2);
+		t5.start((uint8_t)2);
 
 		CommonTestUtils::start<Os>();
